requirements.txt
aiohttp>=3.9.0,<4
requests>=2.31.0,<3
python-dotenv>=1.0.0,<2
py-clob-client>=0.19.0
websockets>=12.0,<14
web3>=6.15.0,<8
fastapi>=0.104.0,<1
uvicorn[standard]>=0.24.0,<1
jinja2>=3.1.0,<4
python-multipart>=0.0.6,<1
pydantic>=2.0.0,<3


